pkg: move InfrastructureConfig to pkg/operator
