+++ pkg/manifests/config_test.go
@@ -295,51 +295,3 @@ func TestLoadProxy(t *testing.T) {
 		})
 	}
 }
-
-func TestNewInfrastructureConfig(t *testing.T) {
-	for _, tc := range []struct {
-		name               string
-		infrastructure     configv1.Infrastructure
-		hostedControlPlane bool
-		haInfrastructure   bool
-	}{
-		{
-			name:               "empty infrastructure",
-			infrastructure:     configv1.Infrastructure{},
-			hostedControlPlane: false,
-			haInfrastructure:   true,
-		},
-		{
-			name: "IBM infrastructure",
-			infrastructure: configv1.Infrastructure{
-				Status: configv1.InfrastructureStatus{
-					Platform: configv1.IBMCloudPlatformType,
-				},
-			},
-			hostedControlPlane: true,
-			haInfrastructure:   true,
-		},
-		{
-			name: "Single-node infrastructure",
-			infrastructure: configv1.Infrastructure{
-				Status: configv1.InfrastructureStatus{
-					InfrastructureTopology: configv1.SingleReplicaTopologyMode,
-				},
-			},
-			hostedControlPlane: false,
-			haInfrastructure:   false,
-		},
-	} {
-		t.Run(tc.name, func(t *testing.T) {
-			c := NewInfrastructureConfig(&tc.infrastructure)
-
-			if c.HostedControlPlane() != tc.hostedControlPlane {
-				t.Errorf("expected hosted control plane: %v, got %v", tc.hostedControlPlane, c.HostedControlPlane())
-			}
-
-			if c.HighlyAvailableInfrastructure() != tc.haInfrastructure {
-				t.Errorf("expected HA infrastructure: %v, got %v", tc.haInfrastructure, c.HighlyAvailableInfrastructure())
-			}
-		})
-	}
-}
